+++ packages/adguard-api/src/content-script/assistant.ts
@@ -15,4 +15,4 @@
  * You should have received a copy of the GNU General Public License
  * along with Adguard API. If not, see <http://www.gnu.org/licenses/>.
  */
-import "@adguard/tswebextension/assistant-inject";
+import '@adguard/tswebextension/assistant-inject';