+++ packages/adguard-api/src/content-script/index.ts
@@ -16,4 +16,4 @@
  * along with Adguard API. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import "@adguard/tswebextension/content-script";
+import '@adguard/tswebextension/content-script';