+++ packages/tswebextension/CHANGELOG.md
@@ -8,6 +8,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version changes -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## [1.0.21] - 2024-03-29
+
+### Fixed
+
+- Do not block "Should collapse" mechanism with `$popup` rules
+
+## [1.0.20] - 2024-03-28
+
+### Changed
+
+- Analysis of $popup rules (in addition to the basic one) to determine
+  the result of blocking a request [AdguardBrowserExtension#2620], [AdguardBrowserExtension#2728].
+- Updated `@adguard/tsurlfilter` to `v2.2.17`.
+- Updated `@adguard/scriptlets` to `v1.10.25`.
+
+[1.0.20]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v1.0.20
+[AdguardBrowserExtension#2620]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2620
+[AdguardBrowserExtension#2728]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2728
+
+
 ## [1.0.18] - 2024-03-25
 
 ### Fixed